Modified to use Arraylist and added status updates
Using an Arryalist is incredibly fast in PowerShell, much better than += when working with big sets of items due to PowerShell's immutable variables.  Also added a Status update as the cmdlet is quiet while doing a lot of work.  Will now display something like this:
 ````
processing rule 103000 of 112270
processing rule 104000 of 112270
processing rule 105000 of 112270
processing rule 106000 of 112270
processing rule 107000 of 112270
````